+++ src/array.js
@@ -58,3 +58,13 @@ export const zipObjectDeepWith = _.curry((x, y) =>
 )
 
 export const flags = zipObjectDeepWith(_, () => true)
+
+export const prefixes = list =>
+  _.range(1, list.length + 1).map(x => _.take(x, list))
+
+export let encoder = separator => ({
+  encode: compactJoin(separator),
+  decode: _.split(separator),
+})
+export let dotEncoder = encoder('.')
+export let slashEncoder = encoder('/')