+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-﻿# 1.40.3
+# 1.41.0
+- Add `pullOn`, `prefixes`
+- Add `encoder`, `dotEncoder`, `slashEncoder`
+- Add `flattenTree`, `treePath`, `propTreePath`, `treeKeys`, `treeValues`
+- Expose `traverse` on `Tree`
+
+# 1.40.3
 - Updated eslint to version 4.16.0.
 
 # 1.40.2